requirements.txt
asyncio==3.4.3
asyncio-nats-client==0.9.2
asyncqt==0.7.0
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
importlib-metadata==0.18
more-itertools==7.2.0
packaging==19.0
pluggy==0.12.0
protobuf==3.9.0
py==1.8.0
pyparsing==2.4.1.1
PyQt5==5.13.0
PyQt5-sip==4.19.18
pytest==5.0.1
pytest-asyncio==0.10.0
six==1.12.0
wcwidth==0.1.7
zipp==0.5.2
